Move IncomesCard, ProjectsCard, SalesCard, CustomerCard and InvoicesCards to their own component.
